+++ src/main/java/at/hannibal2/skyhanni/config/Features.java
@@ -116,6 +116,10 @@ public class Features extends Config {
     public Garden garden = new Garden();
 
     @Expose
+    @Category(name = "Ghost Counter", desc = "Ghost Counter settings.")
+    public GhostCounter ghostCounter = new GhostCounter();
+
+    @Expose
     @Category(name = "Misc", desc = "Settings without a category.")
     public Misc misc = new Misc();
